feat(Feature): Remove redundancy codes
